requirements.txt
# ==============================
# Core scientific stack
# ==============================
numpy==1.24.4
scipy==1.16.2
pandas==2.3.3
matplotlib==3.8.2
Pillow==12.0.0
PyYAML==6.0.3
tqdm==4.67.1
natsort==8.4.0
imageio==2.33.1

# ==============================
# Computer Vision / Geometry
# ==============================
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
trimesh==4.0.8
robust-laplacian==1.0.0
gdist==2.1.0
cholespy==2.1.0

# ==============================
# Deep learning
# ==============================
torch==1.13.1
torchvision==0.14.1
pytorch3d==0.7.3
clip==1.0
scikit-learn==1.7.2

# ==============================
# Blender / Addon specific
# ==============================
# bpy and mathutils are included with Blender; do NOT install via pip
# bpy==3.6.0
# mathutils==3.3.0

# ==============================
# Special / external installation
# ==============================
# nvdiffrast must be installed manually from the repository:
# git clone https://github.com/NVIDIAGameWorks/nvdiffrast

